# docs/openapi/waflo-openapi-current.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user and business memberships.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "user": {
                      "id": "usr_123",
                      "clerkUserId": "user_tavrix_owner",
                      "name": "Tavrix Owner",
                      "email": "owner@tavrix.local",
                      "phone": null,
                      "status": "ACTIVE",
                      "createdAt": "2026-06-10T00:00:00.000Z",
                      "updatedAt": "2026-06-10T00:00:00.000Z"
                    },
                    "memberships": [
                      {
                        "id": "mem_123",
                        "role": "OWNER",
                        "isActive": true,
                        "business": {
                          "id": "bus_123",
                          "name": "Tavrix Cafe",
                          "slug": "tavrix-cafe",
                          "type": "cafe",
                          "city": "Baghdad",
                          "currency": "IQD",
                          "language": "ar",
                          "logoUrl": null,
                          "coverUrl": null
                        }
                      }
                    ],
                    "onboarding": {
                      "hasBusiness": true,
                      "activeBusinessCount": 1,
                      "recommendedNextStep": "OPEN_DASHBOARD"
                    },
                    "businesses": [
                      {
                        "id": "bus_123",
                        "name": "Tavrix Cafe",
                        "slug": "tavrix-cafe",
                        "type": "cafe",
                        "role": "OWNER"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth/me"
        ]
      }
    },
    "/businesses": {
      "post": {
        "operationId": "BusinessesController_createBusiness",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Business created and current user assigned OWNER membership.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "business": {
                      "id": "bus_123",
                      "name": "Tavrix Cafe",
                      "slug": "tavrix-cafe",
                      "type": "cafe",
                      "logoUrl": null,
                      "coverUrl": null,
                      "currency": "IQD",
                      "language": "ar",
                      "city": "Baghdad",
                      "status": "ACTIVE"
                    },
                    "currentMembership": {
                      "id": "mem_123",
                      "role": "OWNER",
                      "isActive": true
                    },
                    "appContext": {
                      "business": {
                        "id": "bus_123",
                        "name": "Tavrix Cafe",
                        "slug": "tavrix-cafe",
                        "type": "cafe",
                        "city": "Baghdad",
                        "currency": "IQD",
                        "language": "ar",
                        "logoUrl": null,
                        "coverUrl": null
                      },
                      "currentMembership": {
                        "id": "mem_123",
                        "role": "OWNER",
                        "isActive": true
                      },
                      "permissions": {
                        "canManageBusiness": true,
                        "canManageMenu": true,
                        "canManageMembers": true,
                        "canViewMembers": true,
                        "canViewPublicLink": true,
                        "canManageAppearance": true
                      },
                      "publicMenu": {
                        "slug": "tavrix-cafe",
                        "path": "/m/tavrix-cafe",
                        "url": "http://localhost:3001/m/tavrix-cafe",
                        "qrPayload": "http://localhost:3001/m/tavrix-cafe"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/me": {
      "get": {
        "operationId": "BusinessesController_getMyBusinesses",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Businesses for the current user.",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "bus_123",
                      "name": "Tavrix Cafe",
                      "slug": "tavrix-cafe",
                      "type": "cafe",
                      "logoUrl": null,
                      "coverUrl": null,
                      "currency": "IQD",
                      "language": "ar",
                      "city": "Baghdad",
                      "status": "ACTIVE",
                      "role": "OWNER"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{id}/app-context": {
      "get": {
        "operationId": "BusinessesController_getAppContext",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Business app context for the Flutter business app.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "business": {
                      "id": "bus_123",
                      "name": "Tavrix Cafe",
                      "slug": "tavrix-cafe",
                      "type": "cafe",
                      "city": "Baghdad",
                      "currency": "IQD",
                      "language": "ar",
                      "logoUrl": null,
                      "coverUrl": null
                    },
                    "currentMembership": {
                      "id": "mem_123",
                      "role": "OWNER",
                      "isActive": true
                    },
                    "permissions": {
                      "canManageBusiness": true,
                      "canManageMenu": true,
                      "canManageMembers": true,
                      "canViewMembers": true,
                      "canViewPublicLink": true,
                      "canManageAppearance": true
                    },
                    "publicMenu": {
                      "slug": "tavrix-cafe",
                      "path": "/m/tavrix-cafe",
                      "url": "http://localhost:3001/m/tavrix-cafe",
                      "qrPayload": "http://localhost:3001/m/tavrix-cafe"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{id}/dashboard-summary": {
      "get": {
        "operationId": "BusinessesController_getDashboardSummary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owner workflow dashboard summary. OWNER, MANAGER, and STAFF can view when actively assigned to this business.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "business": {
                      "id": "bus_123",
                      "name": "Tavrix Cafe",
                      "slug": "tavrix-cafe",
                      "type": "cafe",
                      "city": "Baghdad",
                      "currency": "IQD",
                      "language": "ar",
                      "logoUrl": null,
                      "coverUrl": null
                    },
                    "currentUser": {
                      "role": "OWNER",
                      "permissions": {
                        "canManageBusiness": true,
                        "canManageMenu": true,
                        "canManageMembers": true,
                        "canViewMembers": true,
                        "canViewPublicLink": true,
                        "canManageAppearance": true
                      }
                    },
                    "counts": {
                      "activeCategories": 2,
                      "inactiveCategories": 0,
                      "activeItems": 3,
                      "inactiveItems": 0,
                      "availableItems": 3,
                      "unavailableItems": 0,
                      "activeMembers": 1
                    },
                    "publicMenu": {
                      "path": "/m/tavrix-cafe",
                      "url": "http://localhost:3001/m/tavrix-cafe",
                      "qrPayload": "http://localhost:3001/m/tavrix-cafe"
                    },
                    "onboardingHints": {
                      "hasCategories": true,
                      "hasItems": true,
                      "hasPublicMenuReady": true,
                      "recommendedNextStep": "SHARE_PUBLIC_MENU"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{id}/public-link": {
      "get": {
        "operationId": "BusinessesController_getPublicLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public menu link and QR payload for a business.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "businessId": "bus_123",
                    "slug": "tavrix-cafe",
                    "publicMenuPath": "/m/tavrix-cafe",
                    "publicMenuUrl": "http://localhost:3001/m/tavrix-cafe",
                    "qrPayload": "http://localhost:3001/m/tavrix-cafe"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{id}/menu-appearance": {
      "get": {
        "operationId": "BusinessesController_getMenuAppearance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public menu appearance settings for a business. OWNER, MANAGER, and STAFF can view when actively assigned.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "businessId": "bus_123",
                    "menuTemplateId": "waflo-warm",
                    "menuThemeOverrides": null
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      },
      "patch": {
        "operationId": "BusinessesController_updateMenuAppearance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuAppearanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Public menu appearance updated. OWNER only; STAFF cannot change templates.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "businessId": "bus_123",
                    "menuTemplateId": "coffeehouse-premium",
                    "menuThemeOverrides": null
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{id}/members": {
      "get": {
        "operationId": "BusinessesController_getMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active and inactive business members.",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "mem_123",
                      "userId": "usr_123",
                      "clerkUserId": "user_tavrix_owner",
                      "email": "owner@tavrix.local",
                      "name": "Tavrix Owner",
                      "phone": null,
                      "role": "OWNER",
                      "isActive": true,
                      "createdAt": "2026-06-10T00:00:00.000Z",
                      "updatedAt": "2026-06-10T00:00:00.000Z"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      },
      "post": {
        "operationId": "BusinessesController_createMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Business member created or reactivated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{id}/members/{memberId}": {
      "patch": {
        "operationId": "BusinessesController_updateMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Business member updated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      },
      "delete": {
        "operationId": "BusinessesController_deleteMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Business member deactivated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{id}": {
      "patch": {
        "operationId": "BusinessesController_updateBusiness",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Business profile updated. OWNER only. Supported fields: name, type, city, currency, language, logoUrl, coverUrl."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tavrix Menu API",
    "description": "Tavrix Menu backend API.\n\nProduction and staging clients must send real Clerk JWTs with Authorization: Bearer <clerk-jwt>.\n\nDevelopment can use dev tokens only when NODE_ENV=development.\n\nDevelopment auth header example:\n\nAuthorization: Bearer dev:user_tavrix_owner;email=owner@tavrix.local;name=Tavrix%20Owner",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateBusinessDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Royal Cup"
          },
          "type": {
            "type": "string",
            "example": "cafe"
          },
          "city": {
            "type": "string",
            "example": "Baghdad"
          },
          "currency": {
            "type": "string",
            "example": "IQD",
            "default": "IQD"
          },
          "language": {
            "type": "string",
            "example": "ar",
            "default": "ar"
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "UpdateMenuAppearanceDto": {
        "type": "object",
        "properties": {
          "menuTemplateId": {
            "type": "string",
            "enum": [
              "waflo-warm",
              "coffeehouse-premium",
              "street-bites",
              "minimal-modern"
            ],
            "example": "waflo-warm",
            "description": "Public menu template ID selected for this business."
          },
          "menuThemeOverrides": {
            "type": "object",
            "nullable": true,
            "description": "Reserved for future merchant theme overrides. Public rendering must remain safe when this is null."
          }
        }
      },
      "CreateMemberDto": {
        "type": "object",
        "properties": {
          "clerkUserId": {
            "type": "string",
            "example": "user_tavrix_staff"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "MANAGER",
              "STAFF"
            ],
            "example": "STAFF"
          },
          "email": {
            "type": "string",
            "example": "staff@tavrix.local"
          },
          "name": {
            "type": "string",
            "example": "Tavrix Staff"
          },
          "phone": {
            "type": "string",
            "example": "+9647700000000"
          }
        },
        "required": [
          "clerkUserId",
          "role"
        ]
      },
      "UpdateMemberDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "MANAGER",
              "STAFF"
            ],
            "example": "MANAGER"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "UpdateBusinessDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Royal Cup"
          },
          "type": {
            "type": "string",
            "example": "cafe"
          },
          "city": {
            "type": "string",
            "example": "Baghdad",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "example": "IQD"
          },
          "language": {
            "type": "string",
            "example": "ar"
          },
          "logoUrl": {
            "type": "string",
            "example": "https://example.com/logo.png",
            "nullable": true
          },
          "coverUrl": {
            "type": "string",
            "example": "https://example.com/cover.png",
            "nullable": true
          }
        }
      },
...
    }
  }
}